--- xcm-simulator/src/laosish/configs/mod.rs
@@ -1,3 +1,0 @@
-mod balances;
-mod system;
-mod xcm_config;
