--- xcm-simulator/src/laosish/types/account_id_to_h160.rs
@@ -1,34 +1,0 @@
-// Copyright 2023-2024 Freeverse.io
-// This file is part of LAOS.
-
-// LAOS is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// LAOS is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with LAOS.  If not, see <http://www.gnu.org/licenses/>.
-
-use crate::AccountId;
-use sp_core::H160;
-use sp_runtime::traits::{Convert, ConvertBack};
-
-/// Converts [`AccountId`] to [`H160`]
-pub struct AccountIdToH160;
-
-impl Convert<AccountId, H160> for AccountIdToH160 {
-	fn convert(account_id: AccountId) -> H160 {
-		H160(account_id.0)
-	}
-}
-
-impl ConvertBack<AccountId, H160> for AccountIdToH160 {
-	fn convert_back(account_id: H160) -> AccountId {
-		AccountId::from(account_id)
-	}
-}
