--- xcm-simulator/src/laosish/types/mod.rs
@@ -1,23 +1,0 @@
-// Copyright 2023-2024 Freeverse.io
-// This file is part of LAOS.
-
-// LAOS is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// LAOS is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with LAOS.  If not, see <http://www.gnu.org/licenses/>.
-
-mod account_id_to_h160;
-mod to_author;
-mod transaction_converter;
-
-pub(crate) use account_id_to_h160::AccountIdToH160;
-pub(crate) use to_author::ToAuthor;
-pub use transaction_converter::TransactionConverter;
